==> 2_artemis_explora.ipynb <==
# Importando Biblioteca:

import pandas as pd  # Manipulação de dados tabulares (DataFrames)
import numpy as np  # Manipulação eficiente de arrays e operações matemáticas
import os  # Manipulação de sistema de arquivos

import os
import glob

from google.colab import drive
drive.mount('/content/drive')

# Leitura do Arquivo:

directory = "/content/drive/MyDrive/Exp35-MRO_ contextos_dicas/Exp 35 - rodada 2 - analisados"
addres_file = directory + "/result_artemis/all_data.csv"
df_coords = pd.read_csv(addres_file)
df_coords

# Teste da Função:

def explorar(dataframe, directory, colunas_analisadas=None, group_col=None):
    """
    Explore and analyze the specified columns in a DataFrame, and optionally group data by a specific column.

    Parameters:
    - dataframe (DataFrame): The input DataFrame for exploration.
    - colunas_analisadas (list or None): List of columns to analyze. If None, analyzes all object-type columns.
    - group_col (str or None): Optional column to group data by.

    Returns:
    - None: Writes output to specified file or prints to console.
    """
    import pandas as pd
    from io import StringIO

    output_text = ""  # Inicializar uma string vazia para armazenar a saída

    output_text += f"Resultado da Exploração: Descrição do dataset....\n"
    output_text += "\n"
    output_text += "\n"

    if colunas_analisadas is None:
        colunas_analisadas = dataframe.select_dtypes(include='object').columns

    # Explorar as colunas analisadas
    for coluna in colunas_analisadas:
        valores_unicos = dataframe[coluna].unique()
        output_text += f"Valores únicos na coluna '{coluna}': {valores_unicos}\n"

    if group_col:
        # Agrupar por group_col e imprimir valores únicos de 'id'
        grouped_data = dataframe.groupby(group_col)['id'].unique()
        output_text += f"\nValores únicos de 'id' agrupados por '{group_col}':\n{grouped_data}\n"

    colunas_numericas = dataframe.select_dtypes(include=['float', 'int']).columns
    colunas_likelihood = [col for col in colunas_numericas if col.endswith('likelihood')]

    # Explorar as colunas de likelihood
    for coluna in colunas_likelihood:
        # Calcular as frequências em intervalos de 0.1 e imprimir a tabela de frequências
        intervalos = pd.cut(dataframe[coluna], bins=[0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0])
        frequencias = intervalos.value_counts(normalize=True).sort_index()
        tabela_frequencias = pd.DataFrame({'Valores': frequencias.index, 'Frequência': frequencias.values})
        output_text += f"\nTabela de Frequências para a coluna '{coluna}':\n{tabela_frequencias}\n"

    output_text += "\n"

    dest_directory = f'{directory}/result_artemis'

    if not os.path.exists(dest_directory):
        os.makedirs(dest_directory)

    txt_file_path = f'{dest_directory}/descricao_dataset_coord.txt'

    with open(txt_file_path, 'w') as file:
      file.write(output_text)

    # Confirmar o salvamento
    print(f'O arquivo txt foi salvo no endereço: {txt_file_path}.')

explorar(df_coords, directory=directory, group_col="experimental_context")
