==> 3_artemis_filtra.ipynb <==
def filter_features(dataframe, id=None, phase=None, day=None, session_kind=None, experimental_context=None, context_type=None, sex=None, bodypart=None, selected_columns=None, output_directory=None):
    """
    Applies filters to a DataFrame based on specified values for the given columns and selects specific columns.

    Parameters:
    - dataframe (pd.DataFrame): The target DataFrame.
    - id (str, list, or None): Value or list of values for the 'id' column.
    - phase (str, list, or None): Value or list of values for the 'phase' column.
    - day (int, list, or None): Value or list of values for the 'day' column.
    - session_kind (str, list, or None): Value or list of values for the 'session_kind' column.
    - experimental_context (str, list, or None): Value or list of values for the 'experimental_context' column.
    - context_type (str, list, or None): Value or list of values for the 'context_type' column.
    - sex (str, list, or None): Value or list of values for the 'sex' column.
    - bodypart (str): The body part for which to select columns.
    - selected_columns (list or None): List of columns to select for the final DataFrame.
    - output_directory (str or None): Output directory for saving the result CSV file.

    Returns:
    - pd.DataFrame: Filtered and selected DataFrame.
    """

    import pandas as pd  # Importing pandas at the start of the function

    filters = {
        'id_amostra	': id,
        'rat': id,
        'phase': phase,
        'day': day,
        'session_kind': session_kind,
        'experimental_context': experimental_context,
        'context_type': context_type,
        'sex': sex
    }

    result = dataframe.copy()
    result['frame'] = result.index  # Add 'frame' column with index values

    # Check if filters are provided and filter the DataFrame accordingly
    for column, values in filters.items():
        if values is not None:
            if isinstance(values, list):
                valid_values = [val for val in values if val in dataframe[column].unique()]
                result = result[result[column].isin(valid_values)]
            else:
                if values in dataframe[column].unique():
                    result = result[result[column] == values]

    # Reset 'frame' column to start from zero for each combination of 'id' and 'day'
    result['frame'] = result.groupby(['rat', 'day']).cumcount()

    # Reorganize columns to move 'frame' to the first position
    result = result[['frame'] + [col for col in result.columns if col != 'frame']]

    # Check if selected_columns are provided and select only those columns
    if selected_columns:
        valid_columns = [col for col in selected_columns if col in result.columns]
        selected_dataframe = result[valid_columns]
    else:
        selected_dataframe = result

    # Determine the output directory for saving the CSV file
    if output_directory is not None:
        output_file = os.path.dirname(output_directory)
        output_file = os.path.join(output_file, "dataset_select.csv")
        selected_dataframe.to_csv(output_file, index=False)
        print("the file is saved")

    return selected_dataframe

def filter_likelihood(df, region, min_limit=0.8, output_directory=None):
    """
    Filters out uncertain data points based on the likelihood threshold and performs linear interpolation for missing values.

    Parameters:
    - df (pd.DataFrame): Input DataFrame containing data points and likelihood values.
    - region (str): Name of the region used in the DataFrame, which is part of column names.
    - min_limit (float): Minimum likelihood threshold for filtering out uncertain data points. Default is 6.

    Returns:
    - pd.DataFrame: DataFrame with uncertain data points removed and missing values interpolated.
    """

    import pandas as pd
    import numpy as np

    # Make a copy of the input DataFrame to avoid modifying the original
    df_filtered = df.copy()

    # Filter out data points with likelihood below the minimum limit
    uncertain_data = df_filtered[df_filtered[f'{region}_likelihood'] < min_limit]

    # Replace x and y coordinates with NaN for uncertain data points
    df_filtered.loc[df_filtered[f'{region}_likelihood'] < min_limit, [f'{region}_x', f'{region}_y']] = np.nan

    # Perform linear interpolation to fill missing values
    df_filtered.interpolate(method='linear', inplace=True)

    # Determine the output directory for saving the CSV file
    if output_directory is not None:
        output_file = os.path.dirname(output_directory)
        output_file = os.path.join(output_file, "dataset_filter_likelihood.csv")
        df_filtered.to_csv(output_file, index=False)
        print("the file is saved")

    return uncertain_data, df_filtered

import pandas as pd  # Manipulação de dados tabulares (DataFrames)
import numpy as np  # Manipulação eficiente de arrays e operações matemáticas
import os  # Manipulação de sistema de arquivos

import os
import glob

from google.colab import drive
drive.mount('/content/drive')

directory = "/content/drive/MyDrive/PPGNeuro/area_Exp35/Exp 35 - rodada 2 - analisados/result_artemis/all_data.csv"
df_coords = pd.read_csv(directory)
df_coords

id = "CCM247"
region = "center_b"
day = 1
selected_columns =  ['frame', 'id_amostra', "rat", 'experimental_context',  'day', f'{region}_x', f'{region}_y', f'{region}_likelihood']

df = filter_features(df_coords, bodypart=region, selected_columns=selected_columns, output_directory=directory)
df

uncertain_data, df_end = filter_likelihood(df, region=region, output_directory=directory)
uncertain_data
