==> 4_distancia_velocidade.ipynb <==
# Distance (Space) Functions:

def calcular_diferenca(df, bodypart="center", sig=6):
    """
    Esta função calcula as colunas de diferença (delta_x e delta_y) entre coordenadas consecutivas de uma parte do corpo.

    Parâmetros:
    - df: DataFrame contendo os dados a serem processados.
    - bodypart: Nome da parte do corpo para a qual as coordenadas estão sendo calculadas.

    Retorna:
    - DataFrame com as colunas de diferença adicionadas.
    """
    import pandas as pd
    from scipy.ndimage import gaussian_filter1d

    df_delta = df.copy()

    df_delta['delta_x'] = df[f'{bodypart}_x'].diff()
    df_delta['delta_y'] = df[f'{bodypart}_y'].diff()

    # Preencher os valores NaN com zeros ou valores iniciais
    df_delta['delta_x'].fillna(0, inplace=True)
    df_delta['delta_y'].fillna(0, inplace=True)

    df['distance_x'] = gaussian_filter1d(df_delta['delta_x'], sigma=sig)
    df['distance_y'] = gaussian_filter1d(df_delta['delta_y'], sigma=sig)

    return df

def calcular_distancia_pixels(df):
    """
    Esta função calcula a coluna de distância euclidiana entre pontos consecutivos em um DataFrame.

    Parâmetros:
    - df: DataFrame contendo as coordenadas e as colunas de diferença (delta_x e delta_y).

    Retorna:
    - DataFrame com a coluna de distância adicionada.
    """
    import pandas as pd
    import numpy as np

    df['distance_pixels'] = np.sqrt(df['distance_x']**2 + df['distance_y']**2)
    return df

def calcular_distancia_cm(df, conversion_rate):
    """
    Converts a column in DataFrame from pixels to centimeters.

    Parameters:
    - df (pd.DataFrame): The DataFrame containing the column to be converted.
    - column_name (str): The name of the column to be converted.
    - conversion_rate (float): The conversion rate from pixels to centimeters.

    Returns:
    - pd.DataFrame: The DataFrame with the new column representing distances in centimeters.
    """
    import pandas as pd
    import numpy as np

    df_copy = df.copy()
    df_copy["distance_cm"] = df_copy["distance_pixels"] / conversion_rate
    return df_copy

def spaces(df, conversion_rate):
    import pandas as pd
    import numpy as np

    space_pixels = calcular_distancia_pixels(df)
    result = calcular_distancia_cm(df, conversion_rate)
    return result

# FPS_video:

def calcular_fps(df, min):
    """
    Calcula a taxa de quadros por segundo (FPS) com base na quantidade de quadros e na duração em minutos.

    Parâmetros:
    - df (pd.DataFrame): DataFrame que represente um video
    - min (float): Duração do vídeo em minutos.

    Retorna:
    - float: Taxa de quadros por segundo (FPS).
    """
    fps = len(df)/(min*60)
    return fps

# Time Functions:

def times(df):
    """
    Esta função calcula as colunas de tempo em segundos e minutos com base nos dados do DataFrame e no DataFrame de fps.

    Parâmetros:
    - df: DataFrame contendo os dados a serem processados.
    - fps_df: DataFrame contendo as informações de fps.

    Retorna:
    - DataFrame com as colunas de tempo adicionadas.
    """
    import pandas as pd

    df = df.reset_index(drop=True)

    list_times = []
    result_df = pd.DataFrame()

    # Obter todas as combinações únicas de 'rat' e 'day'
    combinacoes_unicas = df[['rat', 'day']].drop_duplicates()

    # Iterar sobre as combinações únicas
    for _, combinacao in combinacoes_unicas.iterrows():
        rat_unico = combinacao['rat']
        day_unico = combinacao['day']

        temp_df = df[(df['rat'] == rat_unico) & (df['day'] == day_unico)].copy()

        fps_value = 15

        # Calcula as colunas de tempo em segundos e minutos
        temp_df['tempo_segundos'] = temp_df['frame'] / fps_value
        temp_df['tempo_minutos'] = temp_df['tempo_segundos'] / 60

        # Adiciona o resultado ao DataFrame final
        result_df = pd.concat([result_df, temp_df])

    return result_df

# Speed:

def speed_rat(df):
  import pandas as pd

  df = df.reset_index(drop=True)

  list_times = []
  result_df = pd.DataFrame()

  # Obter todas as combinações únicas de 'rat' e 'day'
  combinacoes_unicas = df[['rat', 'day']].drop_duplicates()

  # Iterar sobre as combinações únicas
  for _, combinacao in combinacoes_unicas.iterrows():
      rat_unico = combinacao['rat']
      day_unico = combinacao['day']

      df_speed = df[(df['rat'] == rat_unico) & (df['day'] == day_unico)].copy()

      min = df_speed[f'tempo_minutos'].max()

      fps_value = 15

      df_speed[f'instantaneous_speed'] = df_speed[f'distance_cm']/ (1/fps_value)

      # Adiciona o resultado ao DataFrame final
      result_df = pd.concat([result_df, df_speed])

  return result_df

#  Import Base Libs:

import pandas as pd  # Manipulação de dados tabulares (DataFrames)
import numpy as np  # Manipulação eficiente de arrays e operações matemáticas
import os  # Manipulação de sistema de arquivos

import os
import glob

from google.colab import drive
drive.mount('/content/drive')

# Tests:

directory = "/content/drive/MyDrive/PPGNeuro/area_Exp35/Exp 35 - rodada 2 - analisados/result_artemis/dataset_select.csv"
df_coords = pd.read_csv(directory)
df_coords

regiao = "center_b"
min_likelihood = 0.6
taxa_conversao = 10.91

df = calcular_diferenca(df_coords, regiao)
df

df["day"].unique()

df_result = times(df)
df_result

df_end = spaces(df_result, conversion_rate=taxa_conversao)
df_end

df_ultra = speed_rat(df_end)
df_ultra

# Agrupar por 'rat' e 'day' e calcular a soma das colunas desejadas
result_sum = df_ultra.groupby(['rat', 'day', "experimental_context"])[['distance_pixels', 'distance_cm']].sum().reset_index()

# Agrupar por 'rat' e 'day' e calcular o máximo das colunas desejadas
result_max = df_ultra.groupby(['rat', 'day', "experimental_context"])[['tempo_segundos', 'tempo_minutos']].max().reset_index()

# Mesclar os dois DataFrames com base nas colunas 'rat' e 'day'
result = pd.merge(result_sum, result_max, on=['rat', 'day', "experimental_context"])

result['distance_m'] = result['distance_cm'] / 100

result['velocidade_media'] = result['distance_cm'] / result['tempo_segundos']

# Exibir o resultado
result

directory_out = "/content/drive/MyDrive/PPGNeuro/area_Exp35/Exp 35 - rodada 2 - analisados"
result.to_csv(directory_out+'/result_artemis/distance_speed.csv', index=False)
result.to_excel(directory_out+'/result_artemis/distance_speed.xlsx', index=False)
